Add complete.js file
